完成后台-获取所有权限列表接口

+++ src/main/java/com/tang/travel/mbg/mapper/UmsPermissionMapper.java
@@ -3,13 +3,13 @@ package com.tang.travel.mbg.mapper;
 import com.tang.travel.mbg.model.UmsPermission;
 
 public interface UmsPermissionMapper {
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(UmsPermission record);
 
     int insertSelective(UmsPermission record);
 
-    UmsPermission selectByPrimaryKey(Long id);
+    UmsPermission selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(UmsPermission record);
